+++ common/src/main/java/de/maxhenkel/voicechat/voice/client/ClientManager.java
@@ -18,6 +18,8 @@ import net.minecraft.client.multiplayer.ClientPacketListener;
 import net.minecraft.network.chat.TranslatableComponent;
 
 import javax.annotation.Nullable;
+import java.io.IOException;
+import java.net.InetAddress;
 import java.net.InetSocketAddress;
 import java.net.SocketAddress;
 
@@ -76,15 +78,25 @@ public class ClientManager {
         if (connection != null) {
             try {
                 SocketAddress socketAddress = ClientCompatibilityManager.INSTANCE.getSocketAddress(connection.getConnection());
-                if (socketAddress instanceof InetSocketAddress address) {
-                    client.connect(new InitializationData(address.getHostString(), secretPacket));
-                } else if (socketAddress instanceof LocalAddress) {
-                    client.connect(new InitializationData("127.0.0.1", secretPacket));
-                }
+                client.connect(new InitializationData(resolveAddress(socketAddress), secretPacket));
             } catch (Exception e) {
-                Voicechat.LOGGER.error("Failed to determine server address", e);
+                Voicechat.LOGGER.error("Failed to connect to voice chat server", e);
             }
         }
+    }
+
+    private static String resolveAddress(SocketAddress socketAddress) throws IOException {
+        if (socketAddress instanceof LocalAddress) {
+            return "127.0.0.1";
+        }
+        if (!(socketAddress instanceof InetSocketAddress address)) {
+            throw new IOException(String.format("Failed to determine server address with SocketAddress of type %s", socketAddress.getClass().getSimpleName()));
+        }
+        InetAddress inetAddress = address.getAddress();
+        if (inetAddress == null) {
+            return address.getHostString();
+        }
+        return inetAddress.getHostAddress();
     }
 
     private void onJoinWorld() {